+++ packages/android-app/js/pages/topic.js
@@ -4,12 +4,15 @@ import React from "react";
 import { NativeModules } from "react-native";
 import TopicPage from "@thetimes/topic";
 
-const config = NativeModules.ReactConfig;
-const { track } = NativeModules.ReactAnalytics;
 const { onArticlePress } = NativeModules.TopicEvents;
+const config = NativeModules.ReactConfig;
+const { fetch } = NativeModules.NativeFetch;
+const { track } = NativeModules.ReactAnalytics;
+
+const TopicPageView = TopicPage(config)(fetch);
 
 const View = ({ topicSlug }: { topicSlug: string }) => (
-  <TopicPage
+  <TopicPageView
     topicSlug={topicSlug}
     onArticlePress={onArticlePress}
     analyticsStream={track}