+++ packages/topic/src/topic.js
@@ -1,7 +1,8 @@
 // @flow
 
 import React from "react";
-import { Text } from "react-native";
+import Topic from "@times-components/topic";
+import withClient from "@thetimes/with-client";
 
 type TopicProps = {
   topicSlug: string,
@@ -13,6 +14,18 @@ const TopicPage = ({
   topicSlug,
   onArticlePress,
   analyticsStream
-}: TopicProps) => <Text>Hello {topicSlug}</Text>;
+}: TopicProps) => (
+  <Topic
+    topic={{
+      name: "Chelsea",
+      description: "A swanky part of town."
+    }}
+    slug={"chelsea"}
+    onArticlePress={(event, extras) => onArticlePress(extras.url)}
+    analyticsStream={analyticsStream}
+    page={1}
+    pageSize={5}
+  />
+);
 
-export default TopicPage;
+export default withClient(TopicPage);