+++ packages/ios-app/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.0.3"></a>
+## [0.0.3](https://github.com/newsuk/times-xnative/compare/@thetimes/ios-app@0.0.2...@thetimes/ios-app@0.0.3) (2018-05-21)
+
+
+
+
+**Note:** Version bump only for package @thetimes/ios-app
+
+<a name="0.0.2"></a>
+## [0.0.2](https://github.com/newsuk/times-xnative/compare/@thetimes/ios-app@0.0.1...@thetimes/ios-app@0.0.2) (2018-05-18)
+
+
+
+
+**Note:** Version bump only for package @thetimes/ios-app
+
 <a name="0.0.1"></a>
 ## 0.0.1 (2018-05-15)